refactor: ops metrics standalone
