+++ core/sdk/src/extn/extn_client_message.rs
@@ -58,10 +58,11 @@ use crate::{
             fb_metrics::{BehavioralMetricRequest, MetricsRequest},
             fb_pin::{PinChallengeRequestWithContext, PinChallengeResponse},
             fb_secure_storage::{SecureStorageRequest, SecureStorageResponse},
-            fb_telemetry::{OperationalMetricRequest, TelemetryPayload},
+            fb_telemetry::TelemetryPayload,
         },
         gateway::rpc_gateway_api::RpcRequest,
         manifest::device_manifest::AppLibraryEntry,
+        observability::OperationalMetricRequest,
         protocol::BridgeProtocolRequest,
         pubsub::{PubSubEvents, PubSubRequest, PubSubResponse},
         session::{AccountSessionRequest, AccountSessionResponse, SessionTokenRequest},
