+++ examples/tm_extn/src/tm_ffi.rs
@@ -16,7 +16,7 @@
 //
 
 use ripple_sdk::{
-    api::{firebolt::fb_telemetry::OperationalMetricRequest, status_update::ExtnStatus},
+    api::{observability::OperationalMetricRequest, status_update::ExtnStatus},
     async_channel::Receiver as CReceiver,
     export_channel_builder, export_extn_metadata,
     extn::{
